Issue #21635: Rename Servlet 6.0 Factories bundle
