+++ dev/io.openliberty.servlet.6.0.internal.factories/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>io.openliberty.servlet.6.0.internal.factories</name>
+	<name>io.openliberty.webcontainer.servlet.6.0.internal.factories</name>
 	<comment></comment>
 	<projects>
 	</projects>
